requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
google-generativeai
pytest
pytest-asyncio
httpx
anyio
